Update windows service
Windows services are supported by go with additional syscall
packages. This patch updates the handling of services. Instead
of only register the service it's also needed that the service
responds to special IPC commands. In addition a debug flag is
introduced to run the service in interactive mode. Use for e.g.
`.\myof-service.exe service run --name myof-client --debug`.

Signed-off-by: Tobias Kohlbau <t.kohlbau@myopenfactory.com>