switch from hub.docker.com to ghcr.io
This bundles a couple of changes. Before this CL we tested the
windows client within a docker container. As we stopped supporting
docker on windows quiet a while back it's better to deprecate this
method on moving contrainer registries. This required an overhaul
of the underlying test setup. As a benefit tests are speed up
quiet a lot.

Signed-off-by: Tobias Kohlbau <t.kohlbau@myopenfactory.com>